id refactor int-->long

+++ mavenSSL/src/main/java/us/cryptomaven/controllers/ProductController.java
@@ -18,7 +18,7 @@ public class ProductController {
     }
 
     @RequestMapping("/product/{id}")
-    public String getProductById(@PathVariable Integer id, Model model) {
+    public String getProductById(@PathVariable Long id, Model model) {
         model.addAttribute("product", productService.getProduct(id));
         return "product";
     }
