+++ mavenSSL/src/main/java/us/cryptomaven/repositories/AuthorRepository.java
@@ -3,5 +3,5 @@ package us.cryptomaven.repositories;
 import us.cryptomaven.domain.Author;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorRepository extends CrudRepository<Author, Integer> {
+public interface AuthorRepository extends CrudRepository<Author, Long> {
 }
