+++ mavenSSL/src/main/java/us/cryptomaven/repositories/ProductCategoryRepository.java
@@ -3,5 +3,5 @@ package us.cryptomaven.repositories;
 import us.cryptomaven.domain.ProductCategory;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductCategoryRepository extends CrudRepository<ProductCategory, Integer> {
+public interface ProductCategoryRepository extends CrudRepository<ProductCategory, Long> {
 }
